+++ app/src/main/java/ru/practicum/android/diploma/filter/data/FilterRepositoryImpl.kt
@@ -1,0 +1,5 @@
+package ru.practicum.android.diploma.filter.data
+
+import ru.practicum.android.diploma.filter.domain.api.FilterRepository
+
+class FilterRepositoryImpl : FilterRepository
