+++ app/src/main/java/ru/practicum/android/diploma/filter/domain/api/FilterInteractor.kt
@@ -1,0 +1,3 @@
+package ru.practicum.android.diploma.filter.domain.api
+
+interface FilterInteractor
