+++ app/src/main/java/ru/practicum/android/diploma/filter/domain/impl/FilterInteractorImpl.kt
@@ -1,0 +1,5 @@
+package ru.practicum.android.diploma.filter.domain.impl
+
+import ru.practicum.android.diploma.filter.domain.api.FilterInteractor
+
+class FilterInteractorImpl : FilterInteractor
