+++ app/src/main/java/ru/practicum/android/diploma/filter/ui/country/CountryFragment.kt
@@ -1,4 +1,4 @@
-package ru.practicum.android.diploma
+package ru.practicum.android.diploma.filter.ui.country
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,20 +6,20 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
-import ru.practicum.android.diploma.databinding.FragmentVacancyBinding
+import ru.practicum.android.diploma.databinding.FragmentCountryBinding
 
-class VacancyFragment : Fragment() {
+class CountryFragment : Fragment() {
 
-    private var _binding: FragmentVacancyBinding? = null
+    private var _binding: FragmentCountryBinding? = null
     private val binding get() = _binding!!
-    private val viewModel: VacancyViewModel by viewModels()
+    private val viewModel: CountryViewModel by viewModels()
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = FragmentVacancyBinding.inflate(layoutInflater)
+        _binding = FragmentCountryBinding.inflate(layoutInflater)
         return binding.root
     }
 }
