+++ app/src/main/java/ru/practicum/android/diploma/filter/ui/location/LocationViewModel.kt
@@ -1,4 +1,4 @@
-package ru.practicum.android.diploma
+package ru.practicum.android.diploma.filter.ui.location
 
 import androidx.lifecycle.ViewModel
