+++ app/src/main/java/ru/practicum/android/diploma/filter/ui/region/RegionViewModel.kt
@@ -1,4 +1,4 @@
-package ru.practicum.android.diploma
+package ru.practicum.android.diploma.filter.ui.region
 
 import androidx.lifecycle.ViewModel
