+++ app/src/main/java/ru/practicum/android/diploma/search/data/SearchRepositoryImpl.kt
@@ -1,0 +1,5 @@
+package ru.practicum.android.diploma.search.data
+
+import ru.practicum.android.diploma.search.domain.api.SearchRepository
+
+class SearchRepositoryImpl : SearchRepository
