+++ app/src/main/java/ru/practicum/android/diploma/search/data/network/JobApiService.kt
@@ -1,0 +1,3 @@
+package ru.practicum.android.diploma.search.data.network
+
+interface JobApiService
