+++ app/src/main/java/ru/practicum/android/diploma/search/data/network/NetworkClient.kt
@@ -1,0 +1,3 @@
+package ru.practicum.android.diploma.search.data.network
+
+interface NetworkClient
