+++ app/src/main/java/ru/practicum/android/diploma/search/data/network/RetrofitClient.kt
@@ -1,0 +1,3 @@
+package ru.practicum.android.diploma.search.data.network
+
+class RetrofitClient : NetworkClient
