+++ app/src/main/java/ru/practicum/android/diploma/search/domain/api/SearchInteractor.kt
@@ -1,0 +1,3 @@
+package ru.practicum.android.diploma.search.domain.api
+
+interface SearchInteractor
