+++ app/src/main/java/ru/practicum/android/diploma/search/domain/impl/SearchInteractorImpl.kt
@@ -1,0 +1,5 @@
+package ru.practicum.android.diploma.search.domain.impl
+
+import ru.practicum.android.diploma.search.domain.api.SearchInteractor
+
+class SearchInteractorImpl : SearchInteractor
