+++ app/src/main/java/ru/practicum/android/diploma/search/ui/SearchViewModel.kt
@@ -1,4 +1,4 @@
-package ru.practicum.android.diploma
+package ru.practicum.android.diploma.search.ui
 
 import androidx.lifecycle.ViewModel
