+++ app/src/main/java/ru/practicum/android/diploma/vacancy/data/VacancyRepositoryImpl.kt
@@ -1,0 +1,5 @@
+package ru.practicum.android.diploma.vacancy.data
+
+import ru.practicum.android.diploma.vacancy.domain.api.VacancyRepository
+
+class VacancyRepositoryImpl : VacancyRepository
