+++ app/src/main/java/ru/practicum/android/diploma/vacancy/domain/api/VacancyInteractor.kt
@@ -1,0 +1,3 @@
+package ru.practicum.android.diploma.vacancy.domain.api
+
+interface VacancyInteractor
