+++ app/src/main/java/ru/practicum/android/diploma/vacancy/domain/impl/VacancyInteractorImpl.kt
@@ -1,0 +1,5 @@
+package ru.practicum.android.diploma.vacancy.domain.impl
+
+import ru.practicum.android.diploma.vacancy.domain.api.VacancyInteractor
+
+class VacancyInteractorImpl : VacancyInteractor
