+++ app/src/main/java/ru/practicum/android/diploma/vacancy/ui/VacancyFragment.kt
@@ -1,4 +1,4 @@
-package ru.practicum.android.diploma
+package ru.practicum.android.diploma.vacancy.ui
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,20 +6,20 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
-import ru.practicum.android.diploma.databinding.FragmentSectorBinding
+import ru.practicum.android.diploma.databinding.FragmentVacancyBinding
 
-class SectorFragment : Fragment() {
+class VacancyFragment : Fragment() {
 
-    private var _binding: FragmentSectorBinding? = null
+    private var _binding: FragmentVacancyBinding? = null
     private val binding get() = _binding!!
-    private val viewModel: SectorViewModel by viewModels()
+    private val viewModel: VacancyViewModel by viewModels()
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = FragmentSectorBinding.inflate(layoutInflater)
+        _binding = FragmentVacancyBinding.inflate(layoutInflater)
         return binding.root
     }
 }
