+++ app/src/main/res/layout/fragment_filter.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".FilterFragment">
+    tools:context=".filter.ui.filter.FilterFragment">
 
     <TextView
         android:layout_width="match_parent"
