+++ app/src/main/res/layout/fragment_location.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".LocationFragment">
+    tools:context=".filter.ui.location.LocationFragment">
 
     <TextView
         android:layout_width="match_parent"
