+++ app/src/main/res/layout/fragment_region.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".RegionFragment">
+    tools:context=".filter.ui.region.RegionFragment">
 
     <TextView
         android:layout_width="match_parent"
