+++ app/src/main/res/layout/fragment_search.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SearchFragment">
+    tools:context=".search.ui.SearchFragment">
 
     <TextView
         android:id="@+id/tvSearch"
