+++ app/src/main/res/layout/fragment_sector.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SectorFragment">
+    tools:context=".filter.ui.sector.SectorFragment">
 
     <TextView
         android:layout_width="match_parent"
